test: close coverage gaps for settings and feedback (#134)
- Split monolithic test_api.py (1357 lines) into focused modules
  under the 300-line file-size limit:
  test_auth.py, test_routes.py, test_expenses.py,
  test_feedback.py, test_demo.py
- Add TestSettings class: 8 tests covering auth guard, demo redirect,
  page load, email update (clear/invalid/valid) and persistence
- Add test_feedback_rate_limit: pre-fills feedback_attempts for
  testclient IP and asserts "For mange henvendelser" response
- Add test_feedback_github_api_failure: monkeypatches GITHUB_TOKEN
  and httpx to assert "Kunne ikke sende feedback" on API error
- Fix lint: remove unused assignment (add_response), replace
  list-index-0 pattern with next() in two expense-months tests
